requirements.txt
# Core Dependencies
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0

# Interactive Brokers
ibapi==9.81.1.post1

# Data Processing
pyarrow==12.0.1
msgpack==1.0.5
h5py==3.9.0

# Database
sqlalchemy==2.0.19
alembic==1.11.2
redis==4.6.0

# Web & Dashboard
dash==2.11.1
dash-bootstrap-components==1.4.2
plotly==5.15.0
websocket-client==1.6.1

# Monitoring
prometheus-client==0.17.1
psutil==5.9.5

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.6
colorama==0.4.6
tabulate==0.9.0

# Development (optional - comment out if not needed)
black==23.7.0
isort==5.12.0
pylint==2.17.5
mypy==1.4.1
ipython==8.14.0
jupyter==1.0.0
